Rewrite IP address and port ranges
