+++ .env-template
@@ -4,12 +4,17 @@ DEFGUARD_GATEWAY_SECRET=defguard-gateway-secret
 DEFGUARD_YUBIBRIDGE_SECRET=defguard-yubibridge-secret
 DEFGUARD_SECRET_KEY=defguard-secret-key
 DEFGUARD_URL=http://localhost:8000
+# For localhost only:
+DEFGUARD_COOKIE_INSECURE=true
 # How long auth session lives in seconds
 DEFGUARD_AUTH_SESSION_LIFETIME=604800
 # Optional. Generated based on DEFGUARD_URL if not provided.
 # DEFGUARD_WEBAUTHN_RP_ID=localhost
 DEFGUARD_ADMIN_GROUPNAME=admin
 DEFGUARD_DEFAULT_ADMIN_PASSWORD=pass123
+
+### Logging ###
+DEFGUARD_LOG_LEVEL=info
 
 ### Proxy configuration ###
 # Optional. URL of proxy gRPC server
