+++ Cargo.toml
@@ -17,7 +17,10 @@ axum-client-ip = "0.7"
 axum-extra = { version = "0.10", features = ["cookie-private", "typed-header"] }
 base32 = "0.5"
 base64 = "0.22"
-chrono = { version = "0.4", default-features = false, features = ["clock", "serde"] }
+chrono = { version = "0.4", default-features = false, features = [
+    "clock",
+    "serde",
+] }
 clap = { version = "4.5", features = ["derive", "env"] }
 dotenvy = "0.15"
 humantime = "2.1"
@@ -30,12 +33,14 @@ lettre = { version = "0.11", features = ["tokio1-native-tls"] }
 md4 = "0.10"
 mime_guess = "2.0"
 model_derive = { path = "model-derive" }
-openidconnect = { version = "4.0", default-features = false, optional = true, features = ["reqwest"] }
+openidconnect = { version = "4.0", default-features = false, optional = true, features = [
+    "reqwest",
+] }
 parse_link_header = "0.4"
 paste = "1.0.15"
-pgp = "0.14"
+pgp = "0.15"
 prost = "0.13"
-pulldown-cmark = "0.12"
+pulldown-cmark = "0.13"
 # match version used by sqlx
 rand = "0.8"
 rand_core = { version = "0.6", features = ["getrandom"] }
@@ -52,14 +57,26 @@ serde_json = "1.0"
 serde_urlencoded = "0.7"
 sha-1 = "0.10"
 sha256 = "1.5"
-sqlx = { version = "0.8", features = ["chrono", "ipnetwork", "postgres", "runtime-tokio-native-tls", "uuid"] }
+sqlx = { version = "0.8", features = [
+    "chrono",
+    "ipnetwork",
+    "postgres",
+    "runtime-tokio-native-tls",
+    "uuid",
+] }
 ssh-key = "0.6"
 struct-patch = "0.8"
 tera = "1.20"
 thiserror = "2.0"
 # match axum-extra -> cookies
 time = { version = "0.3", default-features = false }
-tokio = { version = "1", features = ["macros", "parking_lot", "rt-multi-thread", "sync", "time"] }
+tokio = { version = "1", features = [
+    "macros",
+    "parking_lot",
+    "rt-multi-thread",
+    "sync",
+    "time",
+] }
 tokio-stream = "0.1"
 tokio-util = "0.7"
 tonic = { version = "0.12", features = ["gzip", "tls-native-roots"] }
@@ -75,7 +92,9 @@ utoipa = { version = "5", features = ["axum_extras", "chrono", "uuid"] }
 utoipa-swagger-ui = { version = "9", features = ["axum", "vendored"] }
 uuid = { version = "1.9", features = ["v4"] }
 webauthn-authenticator-rs = { version = "0.5" }
-webauthn-rs = { version = "0.5", features = ["danger-allow-state-serialisation"] }
+webauthn-rs = { version = "0.5", features = [
+    "danger-allow-state-serialisation",
+] }
 webauthn-rs-proto = "0.5"
 x25519-dalek = { version = "2.0", features = ["static_secrets"] }
 
@@ -88,7 +107,13 @@ bytes = "1.6"
 claims = "0.8"
 matches = "0.1"
 regex = "1.10"
-reqwest = { version = "0.12", features = ["cookies", "json", "multipart", "rustls-tls", "stream"], default-features = false }
+reqwest = { version = "0.12", features = [
+    "cookies",
+    "json",
+    "multipart",
+    "rustls-tls",
+    "stream",
+], default-features = false }
 serde_qs = "0.13"
 webauthn-authenticator-rs = { version = "0.5", features = ["softpasskey"] }
