+++ src/handlers/mod.rs
@@ -127,10 +127,6 @@ impl From<WebError> for ApiResponse {
                         StatusCode::INTERNAL_SERVER_ERROR,
                     )
                 }
-                AclError::PortOutOfRangeError(port) => ApiResponse::new(
-                    json!({"msg": format!("Port out of range: {port}")}),
-                    StatusCode::UNPROCESSABLE_ENTITY,
-                ),
                 AclError::CannotModifyDeletedRuleError(id) => ApiResponse::new(
                     json!({"msg": format!("Cannot modify deleted ACL rule {id}")}),
                     StatusCode::BAD_REQUEST,
